Move the `shouldTransformTo` method and tests using it to the application tests #16
