+++ src/test/HST/ApplicationTests.hs
@@ -3,7 +3,8 @@
 -- | This module contains basic tests for "HST.Application".
 module HST.ApplicationTests ( testApplication ) where
 
-import           Test.Hspec           ( Spec, context, describe, it )
+import           Polysemy                  ( Members, Sem )
+import           Test.Hspec                ( Spec, context, describe, it )
 
 import           HST.Application           ( processModule )
 import           HST.Effect.Cancel         ( Cancel )
@@ -105,3 +106,39 @@ testProcessModule = context "processModule" $ do
     , "      a4 = undefined"
     , "  in  a3"
     ]
+  context "substitution in modules" $ do
+    it "avoids capture of variables shadowed in let expressions"
+      $ runTest
+      $ let m        = ["module A where", "f (x, y) = let x = y in x"]
+            expected = [ "module A where"
+                       , "f a0 = case a0 of"
+                       , "  (a1, a2) -> let x = a2 in x"
+                       ]
+        in m `shouldTransformTo` expected
+    {- TODO The following test should be used instead of the replacement that
+        is not commented out, after the corresponding issue has been fixed.
+    it "avoids capture of variables shadowed in case expressions" $ runTest $
+      let m        = [ "module A where"
+                     , "f (x, y) = case x of"
+                     , "             y -> (x, y)"]
+          expected = [ "module A where"
+                     , "f a0 = case a0 of"
+                     , "  (a1, a2) -> case a1 of"
+                     , "    y -> (a1, y)"]
+      in m `shouldTransformTo` expected -}
+    it "avoids capture of variables shadowed in case expressions"
+      $ runTest
+      $ let m        = [ "module A where"
+                       , "f (x, y) ="
+                       , "  let r = case x of"
+                       , "            y -> (x, y)"
+                       , "  in  r"
+                       ]
+            expected = [ "module A where"
+                       , "f a1 = case a1 of"
+                       , "  (a2, a3) ->"
+                       , "    let r = case a2 of"
+                       , "              a0 -> (a2, a0)"
+                       , "    in  r"
+                       ]
+        in m `shouldTransformTo` expected
